Refactor vote service
