+++ core/services/votes.py
@@ -1,7 +1,16 @@
+from datetime import timedelta
+
 from django.db import transaction
 from django.db.models import Sum, F
-
-from ..models import Vote, Post, Comment, UserProfile
+from django.utils import timezone
+from ..models import (
+    Vote,
+    Post,
+    Comment,
+    UserProfile,
+    EngagementEvent,
+    PostBurstState,
+)
 from ..ranking import recompute_post_ranks
 
 
@@ -58,6 +67,11 @@ def _cast_vote_once(user, *, post=None, comment=None, want: int) -> int:
             points_cached=F("points_cached") + want
         )
 
+        post_for_metrics = target if target_type == "post" else target.post
+
+    _record_engagement_event(post_for_metrics, user)
+    _update_post_burst_state(post_for_metrics)
+
     try:  # Anti-AstroTurf hook; non-blocking
         from core import anti_astroturf as aa
 
@@ -66,6 +80,68 @@ def _cast_vote_once(user, *, post=None, comment=None, want: int) -> int:
         pass
 
     return total
+
+
+def _record_engagement_event(post, voter):
+    now = timezone.now()
+    age_days = min((now - voter.date_joined).days, 3650)
+    EngagementEvent.objects.create(
+        post=post, event_type="vote", voter_age_days=age_days
+    )
+
+
+def _update_post_burst_state(post):
+    now = timezone.now()
+    state, _ = PostBurstState.objects.get_or_create(
+        post=post,
+        defaults={
+            "window_start": now,
+            "buckets": [0] * 10,
+            "bucket_index": 0,
+            "total_5m": 0,
+        },
+    )
+
+    buckets = list(state.buckets or [])
+    if len(buckets) < 10:
+        buckets += [0] * (10 - len(buckets))
+
+    span = state.bucket_span_seconds
+    index = int((now - state.window_start).total_seconds() // span)
+
+    if index >= 10 or index < 0:
+        if index >= 10:
+            shift = min(index, 10)
+            if shift >= 10:
+                buckets = [0] * 10
+            else:
+                buckets = buckets[shift:] + [0] * shift
+            state.window_start += timedelta(seconds=shift * span)
+        else:
+            buckets = [0] * 10
+            state.window_start = now
+        index = int((now - state.window_start).total_seconds() // span)
+        if index >= 10 or index < 0:
+            buckets = [0] * 10
+            state.window_start = now
+            index = 0
+
+    buckets[index] += 1
+    state.buckets = buckets
+    state.bucket_index = index
+    total = 0
+    for i in range(5):
+        total += buckets[(index - i) % 10]
+    state.total_5m = total
+    state.save(
+        update_fields=[
+            "buckets",
+            "bucket_index",
+            "window_start",
+            "total_5m",
+            "last_updated",
+        ]
+    )
 
 
 def cast_vote_post_once(user, post: Post, want: int) -> int:
